# Improve Performance of ContextVariables