--- a/include/Gaffer/ContextProcessor.h
+++ b/include/Gaffer/ContextProcessor.h
@@ -88,7 +88,7 @@ class IECORE_EXPORT ContextProcessor : public ComputeNode
 		/// Must be implemented to return true if the input is used in `processContext()`.
 		virtual bool affectsContext( const Plug *input ) const = 0;
 		/// Must be implemented to modify context in place.
-		virtual void processContext( Context::EditableScope &context ) const = 0;
+		virtual void processContext( Context::EditableScope &context, IECore::ConstRefCountedPtr &storage ) const = 0;
 
 	private :