This goes on top of #4174

#4174 reduces the performance of ContextVariables slightly, so this looks at ways to speed that back up.

The "Cache variable values for performance" commit is purely internal, and is quite a nice speedup when the ContextVariables is used in many contexts, and the context variables being set don't vary with the context - it saves repeated Data allocs in memberDataAndName, and also repeated string->InternedString conversions.

The last commit is a bit more troublesome, because it requires an API change to processContext, but it is also a noticeable speedup, by avoiding the allocations within Context::setAllocated.  I think that adding the "allocations" parameter kinda makes sense with the new EditableScope API, but see what you think.

Timing results are:
Pre-#4174:  1.48 s
Post-#4174:  1.52 s
Cache Variable Values:  1.10 s
API Change:  0.82 s