@danieldresser-ie 2021-04-14
Forgot to mention:  the final commit requires this Cortex PR: https://github.com/ImageEngine/cortex/pull/1150