@johnhaddon 2021-04-19
LGTM. Marking as blocked until https://github.com/ImageEngine/cortex/pull/1150 has made its way through the system. Also note that we/I still need to update the Changes.md to account for the API change.